// types/schemas/font-sizes.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface FontSizes {
	gameName: number;
	nextGame: number;
	streamName: number;
	donationLink: number;
	timer: number;
	streamRaised: number;
	streamTotal: number;
	teamRaised: number;
	teamTotal: number;
	donations: number;
}